Add LinkedIn to headers.
As requested in issue #1, many people want to have a LinkedIn link in
the header of their resume. This was originally omitted since I had
opted to use a QR code for a LinkedIn profile instead.

This commit adds the `linkedin` option to the `resumeheader` and
`coverletterheader` commands and updates the documentation to show how
to use it.

Closes #1